Renamed branch to conform to naming conventions. Refactored LiftLocalApp so that common code now supports a local application and a production localised application.
