+++ server/main/src/main/resources/main.conf
@@ -50,7 +50,6 @@ akka {
       "akka.tcp://Lift@127.0.0.1:2551",
       "akka.tcp://Lift@127.0.0.1:2552"
     ]
-    jvm-ports = [ 2551, 2552, 2553, 2554 ]
 
     auto-down-unreachable-after = 10s
   }
