+++ server/main/src/main/resources/production.conf
@@ -12,7 +12,6 @@ akka {
 
   cluster {
     seed-nodes = [ ${?SEED_NODES} ]
-    jvm-ports = [ ${akka.remote.netty.tcp.port} ]
     retry = 10s
     min-nr-of-members = 2
   }
@@ -32,6 +31,7 @@ cassandra-snapshot-store {
 }
 
 kafka {
+  # NOTE: we override default Kafka producer properties set in main.conf here
   producer {
     metadata.broker.list = ${?KAFKA_BROKER_LIST}
   }
